#109 detect one parameter
