+++ cloudiscovery/provider/all/data/omitted_resources.py
@@ -1,0 +1,91 @@
+OMITTED_RESOURCES = [
+    "aws_cloudhsm_available_zone",
+    "aws_cloudhsm_hapg",
+    "aws_cloudhsm_hsm",
+    "aws_cloudhsm_luna_client",
+    "aws_dax_default_parameter",
+    "aws_dax_parameter_group",
+    "aws_ec2_reserved_instances_offering",
+    "aws_ec2_snapshot",
+    "aws_ec2_spot_price_history",
+    "aws_ssm_available_patch",
+    "aws_ssm_document",
+    "aws_polly_voice",
+    "aws_lightsail_blueprint",
+    "aws_lightsail_bundle",
+    "aws_lightsail_region",
+    "aws_elastictranscoder_preset",
+    "aws_ec2_vpc_endpoint_service",
+    "aws_dms_endpoint_type",
+    "aws_elasticache_service_update",
+    "aws_elasticache_cache_parameter_group",
+    "aws_rds_source_region",
+    "aws_ssm_association",
+    "aws_ssm_patch_baseline",
+    "aws_ec2_prefix",
+    "aws_ec2_image",
+    "aws_ec2_region",
+    "aws_opsworks_operating_system",
+    "aws_rds_account_attribute",
+    "aws_route53_geo_location",
+    "aws_redshift_cluster_track",
+    "aws_directconnect_location",
+    "aws_dms_account_attribute",
+    "aws_securityhub_standard",
+    "aws_ram_resource_type",
+    "aws_ram_permission",
+    "aws_ec2_account_attribute",
+    "aws_elasticbeanstalk_available_solution_stack",
+    "aws_redshift_account_attribute",
+    "aws_opsworks_user_profile",
+    "aws_directconnect_direct_connect_gateway_association",  # DirectConnect resources endpoint are complicated
+    "aws_directconnect_direct_connect_gateway_attachment",
+    "aws_directconnect_interconnect",
+    "aws_dms_replication_task_assessment_result",
+    "aws_ec2_fpga_image",
+    "aws_ec2_launch_template_version",
+    "aws_ec2_reserved_instancesing",
+    "aws_ec2_spot_datafeed_subscription",
+    "aws_ec2_transit_gateway_multicast_domain",
+    "aws_elasticbeanstalk_configuration_option",
+    "aws_elasticbeanstalk_platform_version",
+    "aws_iam_credential_report",
+    "aws_iam_account_password_policy",
+    "aws_importexport_job",
+    "aws_iot_o_taupdate",
+    "aws_iot_default_authorizer",
+    "aws_workspaces_account",
+    "aws_workspaces_account_modification",
+    "aws_rds_export_task",
+    "aws_rds_custom_availability_zone",
+    "aws_rds_installation_media",
+    "aws_rds_d_bsecurity_group",
+    "aws_translate_text_translation_job",
+    "aws_rekognition_project",
+    "aws_rekognition_stream_processor",
+    "aws_sdb_domain",
+    "aws_redshift_table_restore_status",
+    "aws_iot_v2_logging_level",
+    "aws_license_manager_resource_inventory",
+    "aws_license_manager_license_configuration",
+    "aws_logs_query_definition",
+    "aws_autoscaling_scaling_activity",
+    "aws_cloudwatch_metric",
+    "aws_organizations_handshakes_for_organization",
+    "aws_config_organization_config_rule",
+    "aws_organizations_root",
+    "aws_organizations_delegated_administrator",
+    "aws_organizations_create_account_status",
+    "aws_config_organization_conformance_pack_status",
+    "aws_config_organization_conformance_pack",
+    "aws_ec2_reserved_instances_listing",
+    "aws_redshift_cluster_security_group",
+    "aws_guardduty_organization_admin_account",
+    "aws_elasticache_cache_security_group",
+    "aws_organizations_aws_service_access_for_organization",
+    "aws_organizations_account",
+    "aws_config_organization_config_rule_status",
+    "aws_dynamodb_backup",
+    "aws_ec2_prefix_list",
+    "aws_route53_hosted_zones_by_name",
+]
