+++ cloudiscovery/provider/all/data/on_top_policies.py
@@ -1,0 +1,34 @@
+ON_TOP_POLICIES = [
+    "kafka:ListClusters",
+    "synthetics:DescribeCanaries",
+    "medialive:ListInputs",
+    "cloudhsm:DescribeClusters",
+    "ssm:GetParametersByPath",
+    "servicequotas:Get*",
+    "amplify:ListApps",
+    "autoscaling-plans:DescribeScalingPlans",
+    "medialive:ListChannels",
+    "medialive:ListInputDevices",
+    "mediapackage:ListChannels",
+    "qldb:ListLedgers",
+    "transcribe:ListVocabularies",
+    "glue:GetDatabases",
+    "glue:GetUserDefinedFunctions",
+    "glue:GetSecurityConfigurations",
+    "glue:GetTriggers",
+    "glue:GetCrawlers",
+    "glue:ListWorkflows",
+    "glue:ListMLTransforms",
+    "codeguru-reviewer:ListCodeReviews",
+    "servicediscovery:ListNamespaces",
+    "apigateway:GET",
+    "forecast:ListPredictors",
+    "frauddetector:GetDetectors",
+    "forecast:ListDatasetImportJobs",
+    "frauddetector:GetModels",
+    "frauddetector:GetOutcomes",
+    "networkmanager:DescribeGlobalNetworks",
+    "codeartifact:ListDomains",
+    "ses:GetSendQuota",
+    "codeguru-profiler:ListProfilingGroups",
+]
