+++ cloudiscovery/provider/all/data/required_params_override.py
@@ -1,0 +1,65 @@
+# Trying to fix documentation errors or its lack made by "happy pirates" at AWS
+REQUIRED_PARAMS_OVERRIDE = {
+    "batch": {"ListJobs": ["jobQueue"]},
+    "cloudformation": {
+        "DescribeStackEvents": ["stackName"],
+        "DescribeStackResources": ["stackName"],
+        "GetTemplate": ["stackName"],
+        "ListTypeVersions": ["arn"],
+    },
+    "codecommit": {"GetBranch": ["repositoryName"]},
+    "codedeploy": {
+        "GetDeploymentTarget": ["deploymentId"],
+        "ListDeploymentTargets": ["deploymentId"],
+    },
+    "ecs": {
+        "ListTasks": ["cluster"],
+        "ListServices": ["cluster"],
+        "ListContainerInstances": ["cluster"],
+        "DescribeTasks": ["cluster", "tasks"],
+        "DescribeServices": ["cluster", "services"],
+        "DescribeContainerInstances": ["cluster", "containerInstances"],
+    },
+    "elasticbeanstalk": {
+        "DescribeEnvironmentHealth": ["environmentName"],
+        "DescribeEnvironmentManagedActionHistory": ["environmentName"],
+        "DescribeEnvironmentManagedActions": ["environmentName"],
+        "DescribeEnvironmentResources": ["environmentName"],
+        "DescribeInstancesHealth": ["environmentName"],
+    },
+    "iam": {
+        "GetUser": ["userName"],
+        "ListAccessKeys": ["userName"],
+        "ListServiceSpecificCredentials": ["userName"],
+        "ListSigningCertificates": ["userName"],
+        "ListMFADevices": ["userName"],
+        "ListSSHPublicKeys": ["userName"],
+    },
+    "iot": {"ListAuditFindings": ["taskId"]},
+    "opsworks": {
+        "ListAuditFindings": ["taskId"],
+        "DescribeAgentVersions": ["stackId"],
+        "DescribeApps": ["stackId"],
+        "DescribeCommands": ["deploymentId"],
+        "DescribeDeployments": ["appId"],
+        "DescribeEcsClusters": ["ecsClusterArns"],
+        "DescribeElasticIps": ["stackId"],
+        "DescribeElasticLoadBalancers": ["stackId"],
+        "DescribeInstances": ["stackId"],
+        "DescribeLayers": ["stackId"],
+        "DescribePermissions": ["stackId"],
+        "DescribeRaidArrays": ["stackId"],
+        "DescribeVolumes": ["stackId"],
+    },
+    "ssm": {"DescribeMaintenanceWindowSchedule": ["windowId"],},
+    "shield": {"DescribeProtection": ["protectionId"],},
+    "waf": {
+        "ListActivatedRulesInRuleGroup": ["ruleGroupId"],
+        "ListLoggingConfigurations": ["limit"],
+    },
+    "waf-regional": {
+        "ListActivatedRulesInRuleGroup": ["ruleGroupId"],
+        "ListLoggingConfigurations": ["limit"],
+    },
+    "wafv2": {"ListLoggingConfigurations": ["limit"],},
+}
