+++ cloudiscovery/provider/all/resource/all.py
@@ -1,5 +1,5 @@
 import collections
-import functools
+import itertools
 import re
 from concurrent.futures.thread import ThreadPoolExecutor
 from typing import List, Optional
@@ -8,184 +8,33 @@ from botocore.exceptions import UnknownServiceError
 from botocore.loaders import Loader
 
 from provider.all.command import AllOptions
+from provider.all.data.omitted_resources import OMITTED_RESOURCES
+from provider.all.data.on_top_policies import ON_TOP_POLICIES
+from provider.all.data.required_params_override import REQUIRED_PARAMS_OVERRIDE
+from provider.all.exception import all_exception
 from shared.common import (
     ResourceProvider,
     Resource,
     ResourceDigest,
     message_handler,
     ResourceAvailable,
-    log_critical,
 )
 from shared.common_aws import get_paginator, resource_tags
-
-OMITTED_RESOURCES = [
-    "aws_cloudhsm_available_zone",
-    "aws_cloudhsm_hapg",
-    "aws_cloudhsm_hsm",
-    "aws_cloudhsm_luna_client",
-    "aws_dax_default_parameter",
-    "aws_dax_parameter_group",
-    "aws_ec2_reserved_instances_offering",
-    "aws_ec2_snapshot",
-    "aws_ec2_spot_price_history",
-    "aws_ssm_available_patch",
-    "aws_ssm_document",
-    "aws_polly_voice",
-    "aws_lightsail_blueprint",
-    "aws_lightsail_bundle",
-    "aws_lightsail_region",
-    "aws_elastictranscoder_preset",
-    "aws_ec2_vpc_endpoint_service",
-    "aws_dms_endpoint_type",
-    "aws_elasticache_service_update",
-    "aws_elasticache_cache_parameter_group",
-    "aws_rds_source_region",
-    "aws_ssm_association",
-    "aws_ssm_patch_baseline",
-    "aws_ec2_prefix",
-    "aws_ec2_image",
-    "aws_ec2_region",
-    "aws_opsworks_operating_system",
-    "aws_rds_account_attribute",
-    "aws_route53_geo_location",
-    "aws_redshift_cluster_track",
-    "aws_directconnect_location",
-    "aws_dms_account_attribute",
-    "aws_securityhub_standard",
-    "aws_ram_resource_type",
-    "aws_ram_permission",
-    "aws_ec2_account_attribute",
-    "aws_elasticbeanstalk_available_solution_stack",
-    "aws_redshift_account_attribute",
-    "aws_opsworks_user_profile",
-    "aws_directconnect_direct_connect_gateway_association",  # DirectConnect resources endpoint are complicated
-    "aws_directconnect_direct_connect_gateway_attachment",
-    "aws_directconnect_interconnect",
-    "aws_dms_replication_task_assessment_result",
-    "aws_ec2_fpga_image",
-    "aws_ec2_launch_template_version",
-    "aws_ec2_reserved_instancesing",
-    "aws_ec2_spot_datafeed_subscription",
-    "aws_ec2_transit_gateway_multicast_domain",
-    "aws_elasticbeanstalk_configuration_option",
-    "aws_elasticbeanstalk_platform_version",
-    "aws_iam_credential_report",
-    "aws_iam_account_password_policy",
-    "aws_importexport_job",
-    "aws_iot_o_taupdate",
-    "aws_iot_default_authorizer",
-    "aws_workspaces_account",
-    "aws_workspaces_account_modification",
-    "aws_rds_export_task",
-    "aws_rds_custom_availability_zone",
-    "aws_rds_installation_media",
-    "aws_rds_d_bsecurity_group",
-    "aws_translate_text_translation_job",
-    "aws_rekognition_project",
-    "aws_rekognition_stream_processor",
-    "aws_sdb_domain",
-    "aws_redshift_table_restore_status",
-    "aws_iot_v2_logging_level",
-    "aws_license_manager_resource_inventory",
-    "aws_license_manager_license_configuration",
-    "aws_logs_query_definition",
-    "aws_autoscaling_scaling_activity",
-    "aws_cloudwatch_metric",
-    "aws_organizations_handshakes_for_organization",
-    "aws_config_organization_config_rule",
-    "aws_organizations_root",
-    "aws_organizations_delegated_administrator",
-    "aws_organizations_create_account_status",
-    "aws_config_organization_conformance_pack_status",
-    "aws_config_organization_conformance_pack",
-    "aws_ec2_reserved_instances_listing",
-    "aws_redshift_cluster_security_group",
-    "aws_guardduty_organization_admin_account",
-    "aws_elasticache_cache_security_group",
-    "aws_organizations_aws_service_access_for_organization",
-    "aws_organizations_account",
-    "aws_config_organization_config_rule_status",
-    "aws_dynamodb_backup",
-    "aws_ec2_prefix_list",
-    "aws_route53_hosted_zones_by_name",
-]
-
-# Trying to fix documentation errors or its lack made by "happy pirates" at AWS
-REQUIRED_PARAMS_OVERRIDE = {
-    "batch": {"ListJobs": ["jobQueue"]},
-    "cloudformation": {
-        "DescribeStackEvents": ["stackName"],
-        "DescribeStackResources": ["stackName"],
-        "GetTemplate": ["stackName"],
-        "ListTypeVersions": ["arn"],
-    },
-    "codecommit": {"GetBranch": ["repositoryName"]},
-    "codedeploy": {
-        "GetDeploymentTarget": ["deploymentId"],
-        "ListDeploymentTargets": ["deploymentId"],
-    },
-    "ecs": {
-        "ListTasks": ["cluster"],
-        "ListServices": ["cluster"],
-        "ListContainerInstances": ["cluster"],
-    },
-    "elasticbeanstalk": {
-        "DescribeEnvironmentHealth": ["environmentName"],
-        "DescribeEnvironmentManagedActionHistory": ["environmentName"],
-        "DescribeEnvironmentManagedActions": ["environmentName"],
-        "DescribeEnvironmentResources": ["environmentName"],
-        "DescribeInstancesHealth": ["environmentName"],
-    },
-    "iam": {
-        "GetUser": ["userName"],
-        "ListAccessKeys": ["userName"],
-        "ListServiceSpecificCredentials": ["userName"],
-        "ListSigningCertificates": ["userName"],
-        "ListMFADevices": ["userName"],
-        "ListSSHPublicKeys": ["userName"],
-    },
-    "iot": {"ListAuditFindings": ["taskId"]},
-    "opsworks": {
-        "ListAuditFindings": ["taskId"],
-        "DescribeAgentVersions": ["stackId"],
-        "DescribeApps": ["stackId"],
-        "DescribeCommands": ["deploymentId"],
-        "DescribeDeployments": ["appId"],
-        "DescribeEcsClusters": ["ecsClusterArns"],
-        "DescribeElasticIps": ["stackId"],
-        "DescribeElasticLoadBalancers": ["stackId"],
-        "DescribeInstances": ["stackId"],
-        "DescribeLayers": ["stackId"],
-        "DescribePermissions": ["stackId"],
-        "DescribeRaidArrays": ["stackId"],
-        "DescribeVolumes": ["stackId"],
-    },
-    "ssm": {"DescribeMaintenanceWindowSchedule": ["windowId"],},
-    "shield": {"DescribeProtection": ["protectionId"],},
-    "waf": {
-        "ListActivatedRulesInRuleGroup": ["ruleGroupId"],
-        "ListLoggingConfigurations": ["limit"],
-    },
-    "waf-regional": {
-        "ListActivatedRulesInRuleGroup": ["ruleGroupId"],
-        "ListLoggingConfigurations": ["limit"],
-    },
-    "wafv2": {"ListLoggingConfigurations": ["limit"],},
-}
-
-ON_TOP_POLICIES = [
-    "kafka:ListClusters",
-    "synthetics:DescribeCanaries",
-    "medialive:ListInputs",
-    "cloudhsm:DescribeClusters",
-    "ssm:GetParametersByPath",
-]
 
 SKIPPED_SERVICES = [
     "sagemaker"
 ]  # those services have too unreliable API to make use of it
 
 PARALLEL_SERVICE_CALLS = 80
+
+PLURAL_TO_SINGULAR = {
+    "ies": "y",
+    "status": "status",
+    "ches": "ch",
+    "ses": "s",
+}
+
+LISTING_PREFIXES = ["List", "Get", "Describe"]
 
 
 def _to_snake_case(camel_case):
@@ -216,14 +65,6 @@ def _to_snake_case(camel_case):
     )
 
 
-PLURAL_TO_SINGULAR = {
-    "ies": "y",
-    "status": "status",
-    "ches": "ch",
-    "ses": "s",
-}
-
-
 def singular_from_plural(name: str) -> str:
     if name.endswith("s"):
         for plural_suffix, singular_suffix in PLURAL_TO_SINGULAR.items():
@@ -235,6 +76,14 @@ def singular_from_plural(name: str) -> str:
     return name
 
 
+def is_listing_operation(operation_name: str) -> bool:
+    return (
+        operation_name.startswith("List")
+        or operation_name.startswith("Get")
+        or operation_name.startswith("Describe")
+    )
+
+
 def last_singular_name_element(operation_name):
     last_name = re.findall("[A-Z][^A-Z]*", operation_name)[-1]
     return singular_from_plural(last_name)
@@ -243,7 +92,9 @@ def last_singular_name_element(operation_name):
 def retrieve_resource_name(resource, operation_name):
     resource_name = None
     last_name = last_singular_name_element(operation_name)
-    if "name" in resource:
+    if isinstance(resource, str):
+        resource_name = resource
+    elif "name" in resource:
         resource_name = resource["name"]
     elif "Name" in resource:
         resource_name = resource["Name"]
@@ -260,8 +111,8 @@ def only_one_suffix(resource, suffix):
     id_keys = []
     last_id_val = None
     for key, val in resource.items():
-        if key.lower().endswith(suffix) and not key.lower().endswith(
-            "display" + suffix
+        if key.lower().endswith(suffix.lower()) and not key.lower().endswith(
+            "display" + suffix.lower()
         ):
             id_keys.append(key)
             last_id_val = val
@@ -273,7 +124,9 @@ def only_one_suffix(resource, suffix):
 def retrieve_resource_id(resource, operation_name, resource_name):
     resource_id = resource_name
     last_name = last_singular_name_element(operation_name)
-    if "id" in resource:
+    if isinstance(resource, str):
+        resource_id = resource
+    elif "id" in resource:
         resource_id = resource["id"]
     elif last_name + "Id" in resource:
         resource_id = resource[last_name + "Id"]
@@ -281,8 +134,8 @@ def retrieve_resource_id(resource, operation_name, resource_name):
         resource_id = only_one_suffix(resource, "id")
     elif "arn" in resource:
         resource_id = resource["arn"]
-    elif last_name + "Arn" in resource:
-        resource_id = resource[last_name + "Arn"]
+    elif only_one_suffix(resource, last_name + "arn"):
+        resource_id = only_one_suffix(resource, last_name + "arn")
     elif only_one_suffix(resource, "arn"):
         resource_id = only_one_suffix(resource, "arn")
 
@@ -315,10 +168,11 @@ def operation_allowed(
 def build_resource(
     base_resource, operation_name, resource_type, group
 ) -> Optional[Resource]:
-    if isinstance(base_resource, str):
-        return None
     resource_name = retrieve_resource_name(base_resource, operation_name)
     resource_id = retrieve_resource_id(base_resource, operation_name, resource_name)
+
+    if resource_name is None and resource_id is not None:
+        resource_name = resource_id
 
     if resource_id is None or resource_name is None:
         return None
@@ -330,86 +184,6 @@ def build_resource(
         attributes=attributes,
         tags=resource_tags(base_resource),
     )
-
-
-def all_exception(func):
-    @functools.wraps(func)
-    def wrapper(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        # pylint: disable=broad-except
-        except Exception as e:
-            if func.__qualname__ == "AllResources.analyze_operation":
-                if not args[0].options.verbose:
-                    return
-                exception_str = str(e)
-                if (
-                    "is not subscribed to AWS Security Hub" in exception_str
-                    or "not enabled for securityhub" in exception_str
-                    or "The subscription does not exist" in exception_str
-                    or "calling the DescribeHub operation" in exception_str
-                ):
-                    message_handler(
-                        "Operation {} not accessible, AWS Security Hub is not configured... Skipping".format(
-                            args[2]
-                        ),
-                        "WARNING",
-                    )
-                elif (
-                    "not connect to the endpoint URL" in exception_str
-                    or "not available in this region" in exception_str
-                    or "API is not available" in exception_str
-                ):
-                    message_handler(
-                        "Service {} not available in the selected region... Skipping".format(
-                            args[5]
-                        ),
-                        "WARNING",
-                    )
-                elif (
-                    "Your account is not a member of an organization" in exception_str
-                    or "This action can only be made by accounts in an AWS Organization"
-                    in exception_str
-                    or "The request failed because organization is not in use"
-                    in exception_str
-                ):
-                    message_handler(
-                        "Service {} only available to account in an AWS Organization... Skipping".format(
-                            args[5]
-                        ),
-                        "WARNING",
-                    )
-                elif "is no longer available to new customers" in exception_str:
-                    message_handler(
-                        "Service {} is no longer available to new customers... Skipping".format(
-                            args[5]
-                        ),
-                        "WARNING",
-                    )
-                elif (
-                    "only available to Master account in AWS FM" in exception_str
-                    or "not currently delegated by AWS FM" in exception_str
-                ):
-                    message_handler(
-                        "Operation {} not accessible, not master account in AWS FM... Skipping".format(
-                            args[2]
-                        ),
-                        "WARNING",
-                    )
-                else:
-                    log_critical(
-                        "\nError running operation {}, type {}. Error message {}".format(
-                            args[2], args[1], exception_str
-                        )
-                    )
-            else:
-                log_critical(
-                    "\nError running method {}. Error message {}".format(
-                        func.__qualname__, str(e)
-                    )
-                )
-
-    return wrapper
 
 
 def build_resource_type(aws_service, name):
@@ -425,6 +199,8 @@ def build_resource_type(aws_service, name):
 
 def flatten(d, parent_key="", sep="."):
     items = []
+    if isinstance(d, str):
+        return {}
     for k, v in d.items():
         new_key = parent_key + sep + k if parent_key else k
         if isinstance(v, collections.MutableMapping):
@@ -432,6 +208,42 @@ def flatten(d, parent_key="", sep="."):
         else:
             items.append((new_key, v))
     return dict(items)
+
+
+def operation_required_fields(aws_service, input_model, operation):
+    required_fields = []
+    if "required" in input_model and input_model["required"]:
+        required_fields = input_model["required"]
+    if (
+        aws_service in REQUIRED_PARAMS_OVERRIDE
+        and operation["name"] in REQUIRED_PARAMS_OVERRIDE[aws_service]
+    ):
+        required_fields = REQUIRED_PARAMS_OVERRIDE[aws_service][operation["name"]]
+    return required_fields
+
+
+def get_policy_allowed_calls(iam_client, policy_arn):
+    policy_version_id = iam_client.get_policy(PolicyArn=policy_arn)["Policy"][
+        "DefaultVersionId"
+    ]
+    policy_document = iam_client.get_policy_version(
+        PolicyArn=policy_arn, VersionId=policy_version_id
+    )["PolicyVersion"]["Document"]
+
+    return policy_document
+
+
+def permutate_parameters(operation_parameters):
+    if not operation_parameters:
+        return {}
+    operation_parameter_keys, operation_parameter_values = zip(
+        *operation_parameters.items()
+    )
+    parameters_permutation = [
+        dict(zip(operation_parameter_keys, v))
+        for v in itertools.product(*operation_parameter_values)
+    ]
+    return parameters_permutation
 
 
 class AllResources(ResourceProvider):
@@ -475,6 +287,7 @@ class AllResources(ResourceProvider):
 
         return resources
 
+    # pylint: disable=too-many-locals
     @all_exception
     def analyze_service(self, aws_service, boto_loader, allowed_actions):
         resources = []
@@ -505,41 +318,55 @@ class AllResources(ResourceProvider):
             )
             return None
         for name, operation in service_model["operations"].items():
-            if (
-                name.startswith("List")
-                or name.startswith("Get")
-                or name.startswith("Describe")
-            ):
+            if is_listing_operation(name):
                 has_paginator = name in paginators_model["pagination"]
-                if "input" in operation:
-                    input_model = service_model["shapes"][operation["input"]["shape"]]
-                    if "required" in input_model and input_model["required"]:
-                        continue
-                    if (
-                        aws_service in REQUIRED_PARAMS_OVERRIDE
-                        and operation["name"] in REQUIRED_PARAMS_OVERRIDE[aws_service]
-                    ):
-                        continue
                 resource_type = build_resource_type(aws_service, name)
                 if resource_type in OMITTED_RESOURCES:
                     continue
                 if not operation_allowed(allowed_actions, aws_service, name):
                     continue
-                analyze_operation = self.analyze_operation(
-                    resource_type,
-                    name,
-                    has_paginator,
-                    client,
-                    service_full_name,
-                    aws_service,
-                )
-                if analyze_operation is not None:
-                    resources.extend(analyze_operation)
+
+                operation_parameters = dict()
+                if "input" in operation:
+                    input_model = service_model["shapes"][operation["input"]["shape"]]
+                    required_fields = operation_required_fields(
+                        aws_service, input_model, operation
+                    )
+
+                    if len(required_fields) == 1:
+                        required_field = required_fields[0]
+                        required_field_values = self.check_required_field(
+                            required_field,
+                            service_model,
+                            paginators_model["pagination"],
+                            aws_service,
+                            allowed_actions,
+                            client,
+                            service_full_name,
+                        )
+                        operation_parameters[required_field] = required_field_values
+                    elif len(required_fields) != 0:
+                        continue
+
+                parameters_permutation = permutate_parameters(operation_parameters)
+
+                for parameter_permutation in parameters_permutation:
+                    operation_resources = self.retrieve_operation_resources(
+                        resource_type,
+                        name,
+                        has_paginator,
+                        client,
+                        service_full_name,
+                        aws_service,
+                        parameter_permutation,
+                    )
+                    if operation_resources is not None:
+                        resources.extend(operation_resources)
         return resources
 
     @all_exception
     # pylint: disable=too-many-locals,too-many-arguments
-    def analyze_operation(
+    def retrieve_operation_resources(
         self,
         resource_type,
         operation_name,
@@ -547,7 +374,10 @@ class AllResources(ResourceProvider):
         client,
         service_full_name,
         aws_service,
+        operation_parameters=None,
     ) -> List[Resource]:
+        if operation_parameters is None:
+            operation_parameters = {}
         resources = []
         snake_operation_name = _to_snake_case(operation_name)
         # pylint: disable=too-many-nested-blocks
@@ -556,7 +386,7 @@ class AllResources(ResourceProvider):
                 client=client,
                 operation_name=snake_operation_name,
                 resource_type=resource_type,
-                filters=None,
+                filters=operation_parameters,
             )
             list_metadata = pages.result_keys[0].parsed
             result_key = None
@@ -598,9 +428,8 @@ class AllResources(ResourceProvider):
                     if resource is not None:
                         resources.append(resource)
         else:
-
-            response = getattr(client, snake_operation_name)()
-            for response_elem in response.values():
+            response = getattr(client, snake_operation_name)(**operation_parameters)
+            for response_field, response_elem in response.items():
                 if isinstance(response_elem, list):
                     for response_resource in response_elem:
                         resource = build_resource(
@@ -611,16 +440,22 @@ class AllResources(ResourceProvider):
                         )
                         if resource is not None:
                             resources.append(resource)
+                elif response_field != "ResponseMetadata":
+                    resource = build_resource(
+                        response_elem, operation_name, resource_type, aws_service,
+                    )
+                    if resource is not None:
+                        resources.append(resource)
         return resources
 
     def get_policies_allowed_actions(self):
         if self.options.verbose:
             message_handler("Fetching allowed actions...", "HEADER")
         iam_client = self.options.client("iam")
-        view_only_document = self.get_policy_allowed_calls(
+        view_only_document = get_policy_allowed_calls(
             iam_client, "arn:aws:iam::aws:policy/job-function/ViewOnlyAccess"
         )
-        sec_audit_document = self.get_policy_allowed_calls(
+        sec_audit_document = get_policy_allowed_calls(
             iam_client, "arn:aws:iam::aws:policy/SecurityAudit"
         )
 
@@ -638,12 +473,52 @@ class AllResources(ResourceProvider):
 
         return allowed_actions.keys()
 
-    def get_policy_allowed_calls(self, iam_client, policy_arn):
-        policy_version_id = iam_client.get_policy(PolicyArn=policy_arn)["Policy"][
-            "DefaultVersionId"
-        ]
-        policy_document = iam_client.get_policy_version(
-            PolicyArn=policy_arn, VersionId=policy_version_id
-        )["PolicyVersion"]["Document"]
+    # pylint: disable=too-many-arguments
+    def check_required_field(
+        self,
+        required_field,
+        service_model,
+        pagination_model,
+        aws_service,
+        allowed_actions,
+        client,
+        service_full_name,
+    ):
+        method = None
+        operation = None
+        for operation_name, service_operation in service_model["operations"].items():
+            if operation_name.lower().startswith("list" + required_field[:-1].lower()):
+                method = operation_name
+                operation = service_operation
+        if method is None:
+            return []
 
-        return policy_document
+        has_paginator = method in pagination_model
+        resource_type = build_resource_type(aws_service, method)
+        if resource_type in OMITTED_RESOURCES:
+            return []
+        if not operation_allowed(allowed_actions, aws_service, method):
+            return []
+
+        if "input" in operation:
+            input_model = service_model["shapes"][operation["input"]["shape"]]
+            required_fields = operation_required_fields(
+                aws_service, input_model, operation
+            )
+            if required_fields:
+                return []
+
+        operation_resources = self.retrieve_operation_resources(
+            resource_type,
+            method,
+            has_paginator,
+            client,
+            service_full_name,
+            aws_service,
+        )
+        required_field_ids = []
+
+        for resource in operation_resources:
+            required_field_ids.append(resource.digest.id)
+
+        return required_field_ids
